+++ users/src/sdk/models/shared/organizationaccount.ts
@@ -7,7 +7,7 @@ import {
     OrganizationInheritanceStrategy$,
 } from "./organizationinheritancestrategy";
 import { UserRoles, UserRoles$ } from "./userroles";
-import { z } from "zod";
+import * as z from "zod";
 
 export type OrganizationAccount = {
     billingUserId?: string | null | undefined;