+++ users/RELEASES.md
@@ -29,3 +29,33 @@ Based on:
 - [typescript v2.0.1] users
 ### Releases
 - [NPM v2.0.1] https://www.npmjs.com/package/@wingspan/users/v/2.0.1 - users
+
+## 2024-03-04 01:29:22
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470eea6a7577e003be20a26
+- Speakeasy CLI 1.200.0 (2.277.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.0.2] users
+### Releases
+- [NPM v2.0.2] https://www.npmjs.com/package/@wingspan/users/v/2.0.2 - users
+
+## 2024-03-18 01:27:07
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470eea6a7577e003be20a26
+- Speakeasy CLI 1.209.3 (2.281.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.1.0] users
+### Releases
+- [NPM v2.1.0] https://www.npmjs.com/package/@wingspan/users/v/2.1.0 - users
+
+## 2024-03-25 01:28:16
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470eea6a7577e003be20a26
+- Speakeasy CLI 1.221.0 (2.287.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.1.1] users
+### Releases
+- [NPM v2.1.1] https://www.npmjs.com/package/@wingspan/users/v/2.1.1 - users